bot/requirements.txt
aiogram==3.7.0
SQLAlchemy==2.0.32
aiosqlite==0.20.0
pydantic==2.7.4
python-dotenv==1.0.1
greenlet==3.2.4
